Adding new logic for urlparsing
