+++ tests/data/stackurlhelper/templates/quickstart-compliance-dod-scca/templates/defaults.json
@@ -1,35 +1,34 @@
 [
-  {
-    "ParameterKey": "pDeepSecurityAdminUserName",
-    "ParameterValue": "admin"
-  },
-  {
-    "ParameterKey": "pDBPassword",
-    "ParameterValue": "$[alfred_genpass_28]"
-  },
-  {
-    "ParameterKey": "pEC2KeyPairBastion",
-    "ParameterValue": "cikey"
-  },
-  {
-    "ParameterKey": "pAvailabilityZoneB",
-    "ParameterValue": "$[alfred_gensingleaz_2]"
-  },
-  {
-    "ParameterKey": "pAvailabilityZoneA",
-    "ParameterValue": "$[alfred_gensingleaz_1]"
-  },
-  {
-    "ParameterKey": "pEC2KeyPair",
-    "ParameterValue": "cikey"
-  },
-  {
-    "ParameterKey": "pDeepSecurityAdminPass",
-    "ParameterValue": "$[alfred_genpass_28]"
-  },
-  {
-    "ParameterKey":"pNotifyEmail",
-    "ParameterValue":"distlist@example.org"
-  }
-
+    {
+        "ParameterKey": "pDeepSecurityAdminUserName",
+        "ParameterValue": "admin"
+    },
+    {
+        "ParameterKey": "pDBPassword",
+        "ParameterValue": "$[alfred_genpass_28]"
+    },
+    {
+        "ParameterKey": "pEC2KeyPairBastion",
+        "ParameterValue": "cikey"
+    },
+    {
+        "ParameterKey": "pAvailabilityZoneB",
+        "ParameterValue": "$[alfred_gensingleaz_2]"
+    },
+    {
+        "ParameterKey": "pAvailabilityZoneA",
+        "ParameterValue": "$[alfred_gensingleaz_1]"
+    },
+    {
+        "ParameterKey": "pEC2KeyPair",
+        "ParameterValue": "cikey"
+    },
+    {
+        "ParameterKey": "pDeepSecurityAdminPass",
+        "ParameterValue": "$[alfred_genpass_28]"
+    },
+    {
+        "ParameterKey": "pNotifyEmail",
+        "ParameterValue": "distlist@example.org"
+    }
 ]
