+++ tests/data/stackurlhelper/templates/quickstart-linux-bastion/templates/linux-bastion.template
@@ -1,7 +1,7 @@
 AWSTemplateFormatVersion: 2010-09-09
-Description: LinuxBastion+VPC July,18,2019 QS(0037) (Please do not remove) 
+Description: LinuxBastion+VPC July,18,2019 QS(0037) (Please do not remove)
 Metadata:
-  LICENSE: Apache License, Version 2.0 
+  LICENSE: Apache License, Version 2.0
   'AWS::CloudFormation::Interface':
     ParameterGroups:
       - Label:
@@ -354,32 +354,32 @@ Mappings:
     SUSE-SLES-15-HVM:
       Code: SLES15HVM
 Conditions:
-  2BastionCondition: !Or 
-    - !Equals 
+  2BastionCondition: !Or
+    - !Equals
       - !Ref NumBastionHosts
       - '2'
     - !Condition 3BastionCondition
     - !Condition 4BastionCondition
-  3BastionCondition: !Or 
-    - !Equals 
+  3BastionCondition: !Or
+    - !Equals
       - !Ref NumBastionHosts
       - '3'
     - !Condition 4BastionCondition
-  4BastionCondition: !Equals 
+  4BastionCondition: !Equals
     - !Ref NumBastionHosts
     - '4'
-  GovCloudCondition: !Equals 
+  GovCloudCondition: !Equals
     - !Ref 'AWS::Region'
     - us-gov-west-1
-  UseAlternativeInitialization: !Not 
-    - !Equals 
+  UseAlternativeInitialization: !Not
+    - !Equals
       - !Ref AlternativeInitializationScript
       - ''
-  CreateIAMRole: !Equals 
+  CreateIAMRole: !Equals
     - !Ref AlternativeIAMRole
     - ''
-  UseOSImageOverride: !Not 
-    - !Equals 
+  UseOSImageOverride: !Not
+    - !Equals
       - !Ref OSImageOverride
       - ''
 Resources:
@@ -393,7 +393,7 @@ Resources:
       MetricTransformations:
         - MetricName: SSHCommandCount
           MetricValue: '1'
-          MetricNamespace: !Join 
+          MetricNamespace: !Join
             - /
             - - AWSQuickStart
               - !Ref 'AWS::StackName'
@@ -422,9 +422,9 @@ Resources:
         Statement:
           - Action:
               - 's3:GetObject'
-            Resource: !Sub 
+            Resource: !Sub
               - 'arn:${Partition}:s3:::${QSS3BucketName}/${QSS3KeyPrefix}*'
-              - Partition: !If 
+              - Partition: !If
                   - GovCloudCondition
                   - aws-us-gov
                   - aws
@@ -438,9 +438,9 @@ Resources:
               - 'logs:PutRetentionPolicy'
               - 'logs:PutMetricFilter'
               - 'logs:CreateLogGroup'
-            Resource: !Sub 
+            Resource: !Sub
               - arn:${Partition}:logs:${AWS::Region}:${AWS::AccountId}:log-group:${BastionMainLogGroup}:*
-              - Partition: !If 
+              - Partition: !If
                   - GovCloudCondition
                   - aws-us-gov
                   - aws
@@ -451,7 +451,7 @@ Resources:
             Resource: '*'
             Effect: Allow
       Roles:
-        - !If 
+        - !If
           - CreateIAMRole
           - !Ref BastionHostRole
           - !Ref AlternativeIAMRole
@@ -460,7 +460,7 @@ Resources:
     Type: 'AWS::IAM::InstanceProfile'
     Properties:
       Roles:
-        - !If 
+        - !If
           - CreateIAMRole
           - !Ref BastionHostRole
           - !Ref AlternativeIAMRole
@@ -514,7 +514,7 @@ Resources:
       'AWS::CloudFormation::Authentication':
         S3AccessCreds:
           type: S3
-          roleName: !If 
+          roleName: !If
             - CreateIAMRole
             - !Ref BastionHostRole
             - !Ref AlternativeIAMRole
@@ -524,12 +524,12 @@ Resources:
         config:
           files:
             /tmp/bastion_bootstrap.sh:
-              source: !If 
+              source: !If
                 - UseAlternativeInitialization
                 - !Ref AlternativeInitializationScript
-                - !Sub 
+                - !Sub
                   - https://${QSS3BucketName}.${QSS3Region}.amazonaws.com/${QSS3KeyPrefix}scripts/bastion_bootstrap.sh
-                  - QSS3Region: !If 
+                  - QSS3Region: !If
                       - GovCloudCondition
                       - s3-us-gov-west-1
                       - s3
@@ -539,7 +539,7 @@ Resources:
               authentication: S3AccessCreds
           commands:
             b-bootstrap:
-              command: !Join 
+              command: !Join
                 - ''
                 - - ./tmp/bastion_bootstrap.sh
                   - ' --banner '
@@ -555,13 +555,13 @@ Resources:
       PlacementTenancy: !Ref BastionTenancy
       KeyName: !Ref KeyPairName
       IamInstanceProfile: !Ref BastionHostProfile
-      ImageId: !If 
+      ImageId: !If
         - UseOSImageOverride
         - !Ref OSImageOverride
-        - !FindInMap 
+        - !FindInMap
           - AWSAMIRegionMap
           - !Ref 'AWS::Region'
-          - !FindInMap 
+          - !FindInMap
             - LinuxAMINameMap
             - !Ref BastionAMIOS
             - Code
@@ -575,7 +575,7 @@ Resources:
             VolumeType: gp2
             Encrypted: true
             DeleteOnTermination: true
-      UserData: !Base64 
+      UserData: !Base64
         'Fn::Join':
           - ''
           - - |
@@ -585,11 +585,11 @@ Resources:
             - for e in $(echo "
             - !Ref EnvironmentVariables
             - |
-              " | tr ',' ' '); do 
+              " | tr ',' ' '); do
             - |2
-                export $e 
+                export $e
             - |
-              done 
+              done
             - |
               export PATH=$PATH:/usr/local/bin
             - |
@@ -616,17 +616,17 @@ Resources:
             - EIP_LIST="
             - !Ref EIP1
             - ','
-            - !If 
+            - !If
               - 2BastionCondition
               - !Ref EIP2
               - 'Null'
             - ','
-            - !If 
+            - !If
               - 3BastionCondition
               - !Ref EIP3
               - 'Null'
             - ','
-            - !If 
+            - !If
               - 4BastionCondition
               - !Ref EIP4
               - 'Null'
@@ -703,7 +703,7 @@ Outputs:
       Name: !Sub '${AWS::StackName}-BastionSecurityGroupID'
   BastionHostRole:
     Description: Bastion IAM Role name
-    Value: !If 
+    Value: !If
       - CreateIAMRole
       - !Ref BastionHostRole
       - !Ref AlternativeIAMRole
