+++ tests/data/stackurlhelper/templates/quickstart-microsoft-activedirectory/submodules/quickstart-microsoft-rdgateway/templates/rdgw-standalone.template
