+++ tests/data/stackurlhelper/templates/quickstart-microsoft-activedirectory/templates/Template_1_AD_2012R2.template
