+++ tests/data/stackurlhelper/templates/quickstart-microsoft-activedirectory/templates/ad-2.template
