+++ tests/data/stackurlhelper/templates/quickstart-microsoft-activedirectory/templates/archive/ad-1-old.template
