+++ tests/data/stackurlhelper/templates/quickstart-microsoft-activedirectory/templates/archive/ad-2012r2-1.template
