+++ tests/data/stackurlhelper/templates/quickstart-microsoft-exchange/submodules/quickstart-microsoft-activedirectory/submodules/quickstart-microsoft-rdgateway/templates/rdgw-domain.template
