+++ tests/data/stackurlhelper/templates/quickstart-microsoft-exchange/submodules/quickstart-microsoft-activedirectory/submodules/quickstart-microsoft-rdgateway/templates/rdgw-master.template
@@ -234,7 +234,7 @@
             "RuleCondition": {
                 "Fn::Not": [
                     {
-                        "Fn::Not": 
+                        "Fn::Not":
                         [
                             {
                                 "Fn::Contains": [
