+++ tests/data/stackurlhelper/templates/quickstart-microsoft-exchange/submodules/quickstart-microsoft-activedirectory/templates/ad-1.template
@@ -243,7 +243,7 @@ Conditions:
 Resources:
   QuickStartLogs:
     Type: AWS::Logs::LogGroup
-    Properties: 
+    Properties:
       LogGroupName: !Sub '/aws/Quick_Start/${AWS::StackName}'
       RetentionInDays: 30
   DHCPOptions:
@@ -271,10 +271,10 @@ Resources:
       Content:
         schemaVersion: "0.3"
         description: "Deploy AD with SSM Automation"
-        # Role that is utilized to perform the steps within the Automation Document. In this case to be able to Signal CFN and Describe Instances. 
+        # Role that is utilized to perform the steps within the Automation Document. In this case to be able to Signal CFN and Describe Instances.
         assumeRole: "{{AutomationAssumeRole}}"
         # Gathering parameters needed to configure DCs in the Quick Start
-        parameters: 
+        parameters:
           ADServer1NetBIOSName:
             default: "DC1"
             description: "NetBIOS name of the first Active Directory server (up to 15 characters)"
@@ -304,11 +304,11 @@ Resources:
           RestoreModeSecParamName:
             description: "AWS Secrets Parameter Name for the Restore Mode Password"
             type: "String"
-          DomainDNSName: 
+          DomainDNSName:
             default: "example.com"
             description: "Fully qualified domain name (FQDN) of the forest root domain e.g. example.com"
             type: "String"
-          DomainNetBIOSName: 
+          DomainNetBIOSName:
             default: "example"
             description: "NetBIOS name of the domain (up to 15 characters) for users of earlier versions of Windows e.g. EXAMPLE"
             type: "String"
@@ -337,7 +337,7 @@ Resources:
           inputs:
             Service: ec2
             Api: DescribeInstances
-            Filters:  
+            Filters:
             - Name: "tag:Name"
               Values: [ "{{ADServer1NetBIOSName}}","{{ADServer2NetBIOSName}}" ]
             - Name: "tag:aws:cloudformation:stack-name"
@@ -390,7 +390,7 @@ Resources:
           inputs:
             Service: ec2
             Api: DescribeInstances
-            Filters:  
+            Filters:
             - Name: "tag:Name"
               Values: [ "{{ADServer1NetBIOSName}}" ]
             - Name: "tag:aws:cloudformation:stack-name"
@@ -425,14 +425,14 @@ Resources:
           nextStep: "dc2InstanceId"
           inputs:
             DocumentName: AWS-RunPowerShellScript
-            InstanceIds: 
+            InstanceIds:
               - "{{dc1InstanceId.InstanceId}}"
             CloudWatchOutputConfig:
               CloudWatchOutputEnabled: "true"
               CloudWatchLogGroupName: !Ref 'QuickStartLogs'
             Parameters:
-              commands: 
-                - |     
+              commands:
+                - |
                    function DscStatusCheck () {
                        $LCMState = (Get-DscLocalConfigurationManager).LCMState
                        if ($LCMState -eq 'PendingConfiguration' -Or $LCMState -eq 'PendingReboot') {
@@ -442,9 +442,9 @@ Resources:
                            'Completed'
                        }
                    }
-                   
+
                    Start-DscConfiguration 'C:\AWSQuickstart\ConfigDC1' -Wait -Verbose -Force
-                   
+
                    DscStatusCheck
         # This step grabs the Instance ID for the node that will be configured as the second DC in the new domain.
         - name: "dc2InstanceId"
@@ -454,7 +454,7 @@ Resources:
           inputs:
             Service: ec2
             Api: DescribeInstances
-            Filters:  
+            Filters:
             - Name: "tag:Name"
               Values: [ "{{ADServer2NetBIOSName}}" ]
             - Name: "tag:aws:cloudformation:stack-name"
@@ -489,14 +489,14 @@ Resources:
           nextStep: "DnsConfig"
           inputs:
             DocumentName: AWS-RunPowerShellScript
-            InstanceIds: 
+            InstanceIds:
               - "{{dc2InstanceId.InstanceId}}"
             CloudWatchOutputConfig:
               CloudWatchOutputEnabled: "true"
               CloudWatchLogGroupName: !Ref 'QuickStartLogs'
             Parameters:
-              commands: 
-                - |     
+              commands:
+                - |
                    function DscStatusCheck () {
                        $LCMState = (Get-DscLocalConfigurationManager).LCMState
                        if ($LCMState -eq 'PendingConfiguration' -Or $LCMState -eq 'PendingReboot') {
@@ -506,9 +506,9 @@ Resources:
                            'Completed'
                        }
                    }
-                   
+
                    Start-DscConfiguration 'C:\AWSQuickstart\ConfigDC2' -Wait -Verbose -Force
-                   
+
                    DscStatusCheck
         # Ensure that AD servers point to themselves for DNS
         - name: "DnsConfig"
@@ -532,7 +532,7 @@ Resources:
           inputs:
             Choices:
             - NextStep: signalsuccess
-              Not: 
+              Not:
                 Variable: "{{StackName}}"
                 StringEquals: ""
             - NextStep: sleepend
@@ -634,7 +634,7 @@ Resources:
                 Action:
                   - secretsmanager:GetSecretValue
                   - secretsmanager:DescribeSecret
-                Resource: 
+                Resource:
                   - !Ref 'ADAdminSecrets'
                   - !Ref 'RestoreModeSecrets'
                   - !Ref 'ADAltUserSecrets'
@@ -792,11 +792,11 @@ Resources:
           ToPort: 80
           CidrIp: !Ref VPCCIDR
         - IpProtocol: tcp
-          FromPort: 53 
+          FromPort: 53
           ToPort: 53
           CidrIp: !Ref VPCCIDR
         - IpProtocol: udp
-          FromPort: 53 
+          FromPort: 53
           ToPort: 53
           CidrIp: !Ref VPCCIDR
         - IpProtocol: tcp
@@ -896,20 +896,20 @@ Resources:
           ToPort: 137
           SourceSecurityGroupId: !Ref DomainMembersSG
         - IpProtocol: tcp
-          FromPort: 139 
+          FromPort: 139
           ToPort: 139
-          SourceSecurityGroupId: !Ref DomainMembersSG 
+          SourceSecurityGroupId: !Ref DomainMembersSG
         - IpProtocol: tcp
           FromPort: 5722
           ToPort: 5722
           SourceSecurityGroupId: !Ref DomainMembersSG
         - IpProtocol: icmp
-          FromPort: -1 
-          ToPort: -1 
+          FromPort: -1
+          ToPort: -1
           SourceSecurityGroupId: !Ref DomainMembersSG
-  DomainMembersSG: 
+  DomainMembersSG:
     Type: AWS::EC2::SecurityGroup
-    Properties: 
+    Properties:
       GroupDescription: Domain Members
       VpcId: !Ref VPCID
       SecurityGroupIngress:
@@ -1012,11 +1012,11 @@ Resources:
   DCSecurityGroupIngress:
     Type: AWS::EC2::SecurityGroupIngress
     Properties:
-      Description: Security Group Rule between Domain Controllers 
+      Description: Security Group Rule between Domain Controllers
       GroupId: !Ref DomainControllersSG
-      IpProtocol: '-1' 
-      FromPort: -1 
-      ToPort: -1 
+      IpProtocol: '-1'
+      FromPort: -1
+      ToPort: -1
       SourceSecurityGroupId: !Ref DomainControllersSG
 Outputs:
   DomainAdmin:
