+++ tests/data/stackurlhelper/templates/quickstart-microsoft-exchange/submodules/quickstart-microsoft-activedirectory/templates/ad-3.template
@@ -419,7 +419,7 @@
                 "Ref": "MicrosoftAD"
             },
             "Description": "Directory Services ID"
-        },        
+        },
         "DomainAdmin": {
             "Value": {
                 "Fn::Join": [
