+++ tests/data/stackurlhelper/templates/saas-identity-cognito/templates/config.template
