+++ tests/data/stackurlhelper/templates/saas-identity-cognito/templates/copy-files.template
