+++ tests/test_cfn_template.py
@@ -1,5 +1,4 @@
 import unittest
-import unittest
 from pathlib import Path
 
 from taskcat import Config
